backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
google-generativeai==0.8.3
groq==0.11.0
celery==5.4.0
redis==5.2.1
python-dotenv==1.0.1
pydantic-settings==2.7.0
httpx==0.27.2
psycopg2-binary==2.9.9
pgvector==0.3.2
google-auth==2.38.0
google-auth-oauthlib==1.3.1
